+++ drivers/wifi/nrf71/inc/common/vtf.h
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2026 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @brief Header containing VTF parameters specific declarations
+ */
+
+#ifndef __VTF_H__
+#define __VTF_H__
+
+#include <common/fmac_structs_common.h>
+
+/**
+ * @brief Configure VTF parameters in the RPU.
+ *
+ * @param dev_ctx Pointer to the FMAC device context.
+ * @param voltage Voltage reading used for VTF configuration.
+ * @param temp Temperature reading used for VTF configuration.
+ * @param x0 X0 frequency value used for VTF configuration.
+ * @param vtf_buffer_start_address Output pointer to the allocated VTF parameter buffer.
+ *
+ * @retval NRF_WIFI_STATUS_SUCCESS On success.
+ * @retval NRF_WIFI_STATUS_FAIL On failure.
+ */
+enum nrf_wifi_status nrf_wifi_fmac_config_vtf_params(struct nrf_wifi_fmac_dev_ctx *dev_ctx,
+						     unsigned int voltage,
+						     unsigned int temp,
+						     unsigned int x0,
+						     unsigned int *vtf_buffer_start_address);
+#endif /* __VTF_H__ */
