+++ drivers/wifi/nrf71/inc/radio_test/main.h
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2026 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @brief Header containing radio test main specific declarations
+ */
+
+#ifndef __RADIO_TEST_MAIN_H__
+#define __RADIO_TEST_MAIN_H__
+
+#include <radio_test/api.h>
+
+extern struct nrf_wifi_rt_drv_priv rt_drv_priv;
+
+#endif /* __RADIO_TEST_MAIN_H__ */
