+++ drivers/wifi/nrf71/inc/system/wifi_mgmt_scan.h
@@ -14,7 +14,7 @@
 #include <zephyr/device.h>
 #include <zephyr/net/wifi_mgmt.h>
 
-#include "osal_api.h"
+#include <osal_api.h>
 
 /* Fallback RSSI (dBm) for scan results with no valid signal info
  */
