+++ drivers/wifi/nrf71/inc/system/wifi_util.h
