+++ drivers/wifi/nrf71/src/offload_raw_tx/api.c
@@ -1,0 +1,429 @@
+/*
+ * Copyright (c) 2024 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @brief File containing API definitions for the Offloaded raw TX feature.
+ */
+
+#include <ctype.h>
+#include <stdlib.h>
+#include <string.h>
+#include <zephyr/kernel.h>
+#include <zephyr/logging/log.h>
+#include <zephyr/drivers/wifi/nrf_wifi/off_raw_tx/off_raw_tx_api.h>
+#include <offload_raw_tx/fmac_api.h>
+#include <nrf71_wifi_ctrl.h>
+#include <nrf71_wifi_rf.h>
+#include <util.h>
+#include <offload_raw_tx/api.h>
+
+#define DT_DRV_COMPAT nordic_wlan
+LOG_MODULE_DECLARE(wifi_nrf, CONFIG_WIFI_NRF71_LOG_LEVEL);
+
+extern const struct nrf_wifi_osal_ops nrf_wifi_os_zep_ops;
+struct nrf_wifi_off_raw_tx_drv_priv off_raw_tx_drv_priv;
+
+static const int valid_data_rates[] = { 1, 2, 55, 11, 6, 9, 12, 18, 24, 36, 48, 54,
+				  0, 1, 2, 3, 4, 5, 6, 7, -1 };
+
+#ifdef CONFIG_WIFI_FIXED_MAC_ADDRESS_ENABLED
+static int bytes_from_str(uint8_t *buf, int buf_len, const char *src)
+{
+	size_t i;
+	size_t src_len = strlen(src);
+	char *endptr;
+
+	for (i = 0U; i < src_len; i++) {
+		if (!isxdigit((unsigned char)src[i]) &&
+		    src[i] != ':') {
+			return -EINVAL;
+		}
+	}
+
+	(void)memset(buf, 0, buf_len);
+
+	for (i = 0U; i < (size_t)buf_len; i++) {
+		buf[i] = (uint8_t)strtol(src, &endptr, 16);
+		src = ++endptr;
+	}
+
+	return 0;
+}
+#endif /* CONFIG_WIFI_FIXED_MAC_ADDRESS_ENABLED */
+
+
+int nrf70_off_raw_tx_init(uint8_t *mac_addr, unsigned char *country_code)
+{
+	enum nrf_wifi_status status = NRF_WIFI_STATUS_FAIL;
+	struct nrf_wifi_off_raw_tx_drv_ctx *drv_ctx = NULL;
+	void *rpu_ctx = NULL;
+	k_spinlock_key_t key;
+	unsigned char op_band = nrf_wifi_utils_get_op_band();
+	struct nrf_wifi_tx_pwr_ctrl_params tx_pwr_ctrl_params;
+	struct nrf_wifi_tx_pwr_ceil_params tx_pwr_ceil_params;
+	struct nrf_wifi_board_params board_params;
+	unsigned int fw_ver = 0;
+
+	/* The OSAL layer needs to be initialized before any other initialization
+	 * so that other layers (like FW IF,HW IF etc) have access to OS ops
+	 */
+	nrf_wifi_osal_init(&nrf_wifi_os_zep_ops);
+
+	key = k_spin_lock(&off_raw_tx_drv_priv.lock);
+
+	off_raw_tx_drv_priv.fmac_priv = nrf_wifi_off_raw_tx_fmac_init();
+
+	if (off_raw_tx_drv_priv.fmac_priv == NULL) {
+		LOG_ERR("%s: Failed to initialize nRF70 driver",
+			__func__);
+		goto err;
+	}
+
+	drv_ctx = &off_raw_tx_drv_priv.drv_ctx;
+
+	drv_ctx->drv_priv = &off_raw_tx_drv_priv;
+
+	rpu_ctx = nrf_wifi_off_raw_tx_fmac_dev_add(off_raw_tx_drv_priv.fmac_priv,
+						   drv_ctx);
+	if (!rpu_ctx) {
+		LOG_ERR("%s: Failed to add device", __func__);
+		drv_ctx = NULL;
+		goto err;
+	}
+
+	drv_ctx->rpu_ctx = rpu_ctx;
+
+	status = nrf_wifi_fmac_ver_get(rpu_ctx,
+				       &fw_ver);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: Failed to read the firmware version", __func__);
+		goto err;
+	}
+
+	LOG_DBG("Firmware (v%d.%d.%d.%d) booted successfully",
+		NRF_WIFI_UMAC_VER(fw_ver),
+		NRF_WIFI_UMAC_VER_MAJ(fw_ver),
+		NRF_WIFI_UMAC_VER_MIN(fw_ver),
+		NRF_WIFI_UMAC_VER_EXTRA(fw_ver));
+
+	memset(drv_ctx->phy_rf_params_addr, 0, sizeof(drv_ctx->phy_rf_params_addr));
+	status = nrf_wifi_fmac_config_rf_params(drv_ctx->rpu_ctx, drv_ctx->phy_rf_params_addr);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: Failed to configure RF params", __func__);
+		goto err;
+	}
+
+	/* TODO: Remove hardcodes once we hook in sensor readings */
+	status = nrf_wifi_fmac_config_vtf_params(drv_ctx->rpu_ctx,
+						 243,
+						 25,
+						 0,
+						 &drv_ctx->vtf_buffer_start_address);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: Failed to configure VTF params", __func__);
+		goto err;
+	}
+
+	status = nrf_wifi_off_raw_tx_fmac_dev_init(drv_ctx->rpu_ctx,
+#ifdef CONFIG_NRF_WIFI_LOW_POWER
+						   HW_SLEEP_ENABLE,
+#endif /* CONFIG_NRF_WIFI_LOW_POWER */
+						   NRF_WIFI_DEF_PHY_CALIB,
+						   op_band,
+						   IS_ENABLED(CONFIG_NRF_WIFI_BEAMFORMING),
+						   &tx_pwr_ctrl_params,
+						   &tx_pwr_ceil_params,
+						   &board_params,
+						   country_code,
+						   drv_ctx->phy_rf_params_addr,
+						   drv_ctx->vtf_buffer_start_address);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: Firmware initialization failed", __func__);
+		goto err;
+	}
+
+	if (mac_addr) {
+		memcpy(drv_ctx->mac_addr, mac_addr, 6);
+	} else {
+#ifdef CONFIG_WIFI_FIXED_MAC_ADDRESS_ENABLED
+		int ret = -1;
+
+		ret = bytes_from_str(drv_ctx->mac_addr,
+				     6,
+				     CONFIG_WIFI_FIXED_MAC_ADDRESS);
+		if (ret < 0) {
+			LOG_ERR("%s: Failed to parse MAC address: %s",
+				__func__,
+				CONFIG_WIFI_FIXED_MAC_ADDRESS);
+			goto err;
+		}
+#elif CONFIG_WIFI_OTP_MAC_ADDRESS
+	/* Set dummy MAC address */
+	drv_ctx->mac_addr[0] = 0x00;
+	drv_ctx->mac_addr[1] = 0x00;
+	drv_ctx->mac_addr[2] = 0x5E;
+	drv_ctx->mac_addr[3] = 0x00;
+	drv_ctx->mac_addr[4] = 0x10;
+	drv_ctx->mac_addr[5] = 0x00;
+#endif /* CONFIG_WIFI_FIXED_MAC_ADDRESS_ENABLED */
+
+		if (!nrf_wifi_utils_is_mac_addr_valid(drv_ctx->mac_addr)) {
+			LOG_ERR("%s: Invalid MAC address: %02X:%02X:%02X:%02X:%02X:%02X",
+				__func__,
+				drv_ctx->mac_addr[0],
+				drv_ctx->mac_addr[1],
+				drv_ctx->mac_addr[2],
+				drv_ctx->mac_addr[3],
+				drv_ctx->mac_addr[4],
+				drv_ctx->mac_addr[5]);
+			goto err;
+		}
+	}
+
+	k_spin_unlock(&off_raw_tx_drv_priv.lock, key);
+
+	return 0;
+err:
+	if (drv_ctx->rpu_ctx) {
+		nrf_wifi_fmac_dev_rem(drv_ctx->rpu_ctx);
+		drv_ctx->rpu_ctx = NULL;
+	}
+
+	k_spin_unlock(&off_raw_tx_drv_priv.lock, key);
+	nrf70_off_raw_tx_deinit();
+	return -1;
+}
+
+
+void nrf70_off_raw_tx_deinit(void)
+{
+	k_spinlock_key_t key;
+	struct nrf_wifi_off_raw_tx_drv_ctx *drv_ctx = &off_raw_tx_drv_priv.drv_ctx;
+	int i;
+
+	key = k_spin_lock(&off_raw_tx_drv_priv.lock);
+
+	if (!off_raw_tx_drv_priv.fmac_priv) {
+		k_spin_unlock(&off_raw_tx_drv_priv.lock, key);
+		return;
+	}
+
+	nrf_wifi_fmac_deinit(off_raw_tx_drv_priv.fmac_priv);
+
+	for (i = 0; i < NUM_WIFI_PARAMS; i++) {
+		if (drv_ctx->phy_rf_params_addr[i]) {
+			k_free((void *)drv_ctx->phy_rf_params_addr[i]);
+			drv_ctx->phy_rf_params_addr[i] = 0;
+		}
+	}
+	if (drv_ctx->vtf_buffer_start_address) {
+		k_free((void *)drv_ctx->vtf_buffer_start_address);
+		drv_ctx->vtf_buffer_start_address = 0;
+	}
+
+	k_spin_unlock(&off_raw_tx_drv_priv.lock, key);
+}
+
+static bool validate_rate(enum nrf_wifi_off_raw_tx_tput_mode tput_mode,
+			enum nrf_wifi_off_raw_tx_rate rate)
+{
+	if (tput_mode == TPUT_MODE_LEGACY) {
+		if (rate > RATE_54M) {
+			return false;
+		}
+	} else {
+		if (rate <= RATE_54M) {
+			return false;
+		}
+	}
+
+	return true;
+}
+
+int nrf70_off_raw_tx_conf_update(struct nrf_wifi_off_raw_tx_conf *conf)
+{
+	int ret = -1;
+	enum nrf_wifi_status status = NRF_WIFI_STATUS_FAIL;
+	struct nrf_wifi_off_raw_tx_drv_ctx *drv_ctx = &off_raw_tx_drv_priv.drv_ctx;
+	struct nrf_wifi_offload_ctrl_params *off_ctrl_params = NULL;
+	struct nrf_wifi_offload_tx_ctrl *off_tx_params = NULL;
+	struct nrf_wifi_fmac_dev_ctx *fmac_dev_ctx = NULL;
+	k_spinlock_key_t key;
+
+	if (!conf) {
+		LOG_ERR("%s: Config params is NULL", __func__);
+		goto out;
+	}
+
+	off_ctrl_params = nrf_wifi_osal_mem_zalloc(sizeof(*off_ctrl_params));
+	if (!off_ctrl_params) {
+		LOG_ERR("%s: Failed to allocate memory for off_ctrl_params", __func__);
+		goto out;
+	}
+
+	key = k_spin_lock(&off_raw_tx_drv_priv.lock);
+
+	fmac_dev_ctx = drv_ctx->rpu_ctx;
+
+	if (!fmac_dev_ctx) {
+		LOG_ERR("%s: FMAC device context is NULL", __func__);
+		goto out;
+	}
+
+	off_tx_params = nrf_wifi_osal_mem_zalloc(sizeof(*off_tx_params));
+	if (!off_tx_params) {
+		LOG_ERR("%s Failed to allocate memory for off_tx_params: ", __func__);
+		goto out;
+	}
+
+	if (!validate_rate(conf->tput_mode, conf->rate)) {
+		LOG_ERR("%s Invalid rate. Throughput mode: %d, rate: %d\n", __func__,
+				      conf->tput_mode, conf->rate);
+		goto out;
+	}
+
+	off_ctrl_params->chan.primary_num = conf->chan;
+	off_ctrl_params->chan.op_band = nrf_wifi_utils_get_op_band();
+	off_ctrl_params->chan.bw = RPU_CH_BW_20;
+	off_ctrl_params->chan.sec_20_offset = 0;
+	off_ctrl_params->chan.sec_40_offset = 0;
+	off_ctrl_params->period_in_us = conf->period_us;
+	off_ctrl_params->tx_pwr = conf->tx_pwr;
+	off_tx_params->he_gi_type = conf->he_gi;
+	off_tx_params->he_ltf = conf->he_ltf;
+	off_tx_params->pkt_ram_ptr = (unsigned int)conf->pkt;
+	off_tx_params->pkt_length = conf->pkt_len;
+	off_tx_params->rate_flags = conf->tput_mode;
+	off_tx_params->rate = valid_data_rates[conf->rate];
+	off_tx_params->rate_preamble_type = conf->short_preamble;
+	off_tx_params->rate_retries = conf->num_retries;
+
+	status = nrf_wifi_off_raw_tx_fmac_conf(fmac_dev_ctx,
+					       off_ctrl_params,
+					       off_tx_params);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: nRF70 offloaded raw TX configuration failed",
+				      __func__);
+		goto out;
+	}
+
+	ret = 0;
+out:
+	nrf_wifi_osal_mem_free(off_ctrl_params);
+	nrf_wifi_osal_mem_free(off_tx_params);
+	k_spin_unlock(&off_raw_tx_drv_priv.lock, key);
+	return ret;
+}
+
+
+int nrf70_off_raw_tx_start(struct nrf_wifi_off_raw_tx_conf *conf)
+{
+	int ret = -1;
+	enum nrf_wifi_status status = NRF_WIFI_STATUS_FAIL;
+	struct nrf_wifi_off_raw_tx_drv_ctx *drv_ctx = &off_raw_tx_drv_priv.drv_ctx;
+	k_spinlock_key_t key;
+
+	status = nrf70_off_raw_tx_conf_update(conf);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: nRF70 offloaded raw TX configuration failed",
+				      __func__);
+		goto out;
+	}
+
+	key = k_spin_lock(&off_raw_tx_drv_priv.lock);
+	if (!drv_ctx->rpu_ctx) {
+		LOG_ERR("%s: FMAC device context is NULL", __func__);
+		goto out;
+	}
+
+	status = nrf_wifi_off_raw_tx_fmac_start(drv_ctx->rpu_ctx);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: nRF70 offloaded raw TX start failed",
+				      __func__);
+		goto out;
+	}
+
+	ret = 0;
+out:
+	k_spin_unlock(&off_raw_tx_drv_priv.lock, key);
+	return ret;
+}
+
+
+int nrf70_off_raw_tx_stop(void)
+{
+	int ret = -1;
+	enum nrf_wifi_status status = NRF_WIFI_STATUS_FAIL;
+	struct nrf_wifi_off_raw_tx_drv_ctx *drv_ctx = &off_raw_tx_drv_priv.drv_ctx;
+	k_spinlock_key_t key;
+
+	key = k_spin_lock(&off_raw_tx_drv_priv.lock);
+
+	if (!drv_ctx->rpu_ctx) {
+		LOG_ERR("%s: FMAC device context is NULL", __func__);
+		goto out;
+	}
+
+	status = nrf_wifi_off_raw_tx_fmac_stop(drv_ctx->rpu_ctx);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: nRF70 offloaded raw TX stop failed",
+				      __func__);
+		goto out;
+	}
+
+	ret = 0;
+out:
+	k_spin_unlock(&off_raw_tx_drv_priv.lock, key);
+	return ret;
+}
+
+
+int nrf70_off_raw_tx_mac_addr_get(uint8_t *mac_addr)
+{
+	struct nrf_wifi_off_raw_tx_drv_ctx *drv_ctx = &off_raw_tx_drv_priv.drv_ctx;
+
+	if (!mac_addr) {
+		LOG_ERR("%s: Invalid param", __func__);
+		return -EINVAL;
+	}
+
+	memcpy(mac_addr, drv_ctx->mac_addr, 6);
+	return 0;
+}
+
+int nrf70_off_raw_tx_stats(struct nrf_wifi_off_raw_tx_stats *off_raw_tx_stats)
+{
+	int ret = -1;
+	enum nrf_wifi_status status = NRF_WIFI_STATUS_FAIL;
+	struct nrf_wifi_off_raw_tx_drv_ctx *drv_ctx = &off_raw_tx_drv_priv.drv_ctx;
+	struct rpu_off_raw_tx_op_stats stats;
+	k_spinlock_key_t key;
+
+	memset(&stats, 0, sizeof(stats));
+
+	key = k_spin_lock(&off_raw_tx_drv_priv.lock);
+
+	if (!drv_ctx->rpu_ctx) {
+		LOG_ERR("%s: FMAC device context is NULL", __func__);
+		goto out;
+	}
+
+	status = nrf_wifi_off_raw_tx_fmac_stats_get(drv_ctx->rpu_ctx,
+						    0,
+						    &stats);
+	if (status != NRF_WIFI_STATUS_SUCCESS) {
+		LOG_ERR("%s: nRF70 offloaded raw TX stats failed",
+				      __func__);
+		goto out;
+	}
+
+	off_raw_tx_stats->off_raw_tx_pkt_sent = stats.fw.offload_raw_tx_cnt;
+
+	ret = 0;
+out:
+	k_spin_unlock(&off_raw_tx_drv_priv.lock, key);
+	return ret;
+}
